% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_tweets.R
\name{complete_setup}
\alias{complete_setup}
\title{complete_setup}
\usage{
complete_setup()
}
\description{
This function should be called after package installation to properly set up dependencies and create file caching system.
}
