% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_embeddings.R
\name{prepare_glove_embeddings}
\alias{prepare_glove_embeddings}
\title{prepare_glove_embeddings}
\usage{
prepare_glove_embeddings(embedding_dim, tokenizer)
}
\arguments{
\item{embedding_dim}{Dimensionality of word embeddings. Options are 25, 50, 100, 200.}

\item{tokenizer}{Pre-fit keras text tokenizer.}
}
\description{
This function prepares an embedding matrix containing the words in the training data set from pre-trained GloVe embeddings.
}
\details{
For more information on the GloVe embedding algorithm, visit https://nlp.stanford.edu/projects/glove/.
}
\note{
The GloVe embeddings are 1.3G zipped and 3.8G unzipped.

Embeddings are saved as Rdata to a folder called embeddings with the file format "tweet_glove_\{embedding_dim\}.rda"
}
