% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_tweets.R
\name{scrape_tweets}
\alias{scrape_tweets}
\title{scrape_tweets}
\usage{
scrape_tweets(screen_names = NULL, ids = NULL, tweets_per_user,
  credentials_dir, out_dir)
}
\arguments{
\item{screen_names}{Character vector of screen names of Twitter users.}

\item{ids}{Character or integer vector of IDs of Twitter users. Use either (but not both) of these two arguments.}

\item{tweets_per_user}{Number of tweets to scrape for each user.}

\item{credentials_dir}{Directory with Twitter OAuth tokens.}

\item{out_dir}{Name of directory to store scraped Tweets.}
}
\description{
This function scrapes the most recent n Tweets of a list of Twitter users.
}
\examples{
data("tweets")
users <- unique(tweets$screen_name)
scrape_tweets(screen_names = users, tweets_per_user = 200, credentials_dir = "credentials", out_dir = "data/scraped_tweets")
}
