% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_tweets.R
\name{setup_dependencies}
\alias{setup_dependencies}
\title{setup_dependencies}
\usage{
setup_dependencies()
}
\description{
This function should be called after package installation to properly set up dependencies.
}
