% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{texts_to_vectors}
\alias{texts_to_vectors}
\title{texts_to_vectors}
\usage{
texts_to_vectors(texts, tokenizer)
}
\arguments{
\item{texts}{Character vector of raw text data}

\item{tokenizer}{Pre-fit keras tokenizer}
}
\value{
matrix of vectorized texts
}
\description{
Helper function vectorize text data
}
