% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{train_test_split}
\alias{train_test_split}
\title{train_test_split}
\usage{
train_test_split(X, y, test_size = 0.2)
}
\arguments{
\item{X}{data.frame or matrix of data}

\item{y}{Labels (optional).}

\item{test_size}{Proportion of samples to set aside for testing.}
}
\value{
List of X_train, X_test, y_train, y_test
}
\description{
Helper function to split data into training and testing sets.
}
