% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_tweets.R
\name{tweets_to_df}
\alias{tweets_to_df}
\title{tweets_to_df}
\usage{
tweets_to_df(tweet_dir, keep_retweets = FALSE)
}
\arguments{
\item{tweet_dir}{Directory where scraped Tweets are stored}

\item{keep_retweets}{Optionally discard retweets.}
}
\value{
data.frame of Tweets with metadata
}
\description{
This function takes a directory of JSON files containing scraped Tweets and returns a data.frame
}
\examples{
tweet_df <- tweets_to_df("data/scraped_tweets", keep_retweets = FALSE)
}
