% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_ideology}
\alias{predict_ideology}
\title{predict_ideology}
\usage{
predict_ideology(tweets, model = "LSTM", embeddings = "w2v",
  embedding_dim = 25, filter_political_tweets = FALSE)
}
\arguments{
\item{tweets}{Character vector of Tweets.}

\item{model}{Neural network architecture to use. Options are "LSTM", "BiLSTM", or "C-BiLSTM".}

\item{embeddings}{Type of word embedding algorithm to use. Options are "w2v" (word2vec), "glove", or "random" (random initialization).}

\item{embedding_dim}{Length of word embeddings to use. Options are 25, 50, 100, or 200.}

\item{filter_political_tweets}{If Tweet collection may contain non-political Tweets, optionally filter them out before ideological scaling.}
}
\value{
Vector of float values between 0 and 1, where values closer to 0 indiciate liberal ideological slant, values closer to 1 indicate conservative ideological slant, and values near 0.5 indicate a lack of ideological leaning.
}
\description{
This function allows you to scale the ideological slant of Twitter posts.
}
\examples{
tweets <- c("Make no mistake- the President of the United States is actively sabotaging the health insurance of millions of Americans with this action.",
            "This MLK Day, 50 years after his death, we honor Dr. King's legacy. He lived for the causes of justice and equality, and opened the door of opportunity for millions of Americans. America is a better; freer nation because of it.",
            "I’m disappointed in Senate Democrats for shutting down the government. #SchumerShutdown")
}
